Ratelimit get_post endpoint with IP bucket
The get_post endpoint is currently available freely without having to be
authenticated, however this endpoint might be performing a lookup in an
external API for uncached entries.

For that reason, to avoid the possibility of us spamming this external
API, we impose a relatively high rate-limit on it, that is IP address
specific.

Note that doing this may cause issues to bigger networks, such as
schools, which might have a lot of devices connected to a single IP. It
might be worth it to change this endpoint to be a member endpoint
instead, avoiding this issue entirely.