Don't inform users about IP wide rate-limits
By default rate-limit buckets add certain headers to the responses that
inform the client about how many requests they still have in given
period. In most cases this is a good thing as clients can build software
that respects these headers and doesn't get into cooldowns.

However with an IP wide bucket, we don't necessarily expect there to
just be a single computer on that network, meaning showing these headers
wouldn't really make sense, as other people might use up the requests
too. Usually, we shouldn't see the IP rate limit getting crossed anyway
and it's just here to protect us from hitting rate-limits on the the
external API that we rely on.

If a cooldown is triggered though, a special cooldown header will
appear, so that the clients can await it.